requirements.txt
# API clients
requests>=2.25.0
amadeus>=7.0.0

# Web scraping (for future Selenium support)
selenium>=3.141.0
webdriver-manager>=3.5.0

# Notifications
twilio>=7.0.0
python-dotenv>=0.19.0

# Configuration
pyyaml>=5.4.0

# Data handling
pandas>=1.1.0

# Scheduling
schedule>=1.1.0

# Utilities
python-dateutil>=2.8.0
